Updated README.md for Server
Added some more detailed steps for installing ruby and rails and downloading the git repo. 
I did not use rbenv as my ruby manger, so hopefully what remains in the file regarding rbenv still makes sense. 
I got rid of the rbenv rehash command because when I looked it up to find out what it did, I saw it is a deprecated command.