Update README.md
Updated readme on front page to direct people to either frontend or backend, as discussed during our last meeting, there is a lot of information in both of these areas, just not on the front page. Also wanted to point out where you should go if you want to use react.